requirements.txt
selenium
pytest
pytest-quickcheck
pytest-runner
pytest-allure-adaptor
allure-pytest
pyvirtualdisplay
